smaller adjustments to osgEarth Code fixes linux build error
git-svn-id: file:///local/work/git-test/fastprotect.svn/trunk@7553 eaa77fc2-5f71-4410-aed7-0df95d92e94a

+++ src/osgEarth/CacheSeed.cpp
@@ -80,7 +80,7 @@ void CacheSeed::seed( Map* map )
         {
             hasCaches = true;
 
-			if (opt.minLevel().isSet() && opt.minLevel().get() < src_min_level)
+			if (opt.minLevel().isSet() && opt.minLevel().get() < (int)src_min_level)
                 src_min_level = opt.minLevel().get();
 			if (opt.maxLevel().isSet() && opt.maxLevel().get() > (int)src_max_level)
                 src_max_level = opt.maxLevel().get();
@@ -113,7 +113,7 @@ void CacheSeed::seed( Map* map )
         {
             hasCaches = true;
 
-			if (opt.minLevel().isSet() && opt.minLevel().get() < src_min_level)
+			if (opt.minLevel().isSet() && opt.minLevel().get() < (int)src_min_level)
                 src_min_level = opt.minLevel().get();
 			if (opt.maxLevel().isSet() && opt.maxLevel().get() > (int)src_max_level)
                 src_max_level = opt.maxLevel().get();
